WIP swerve module control
Lots of changes to the swerve module to make set_desired_state() _hopefully_ work as desired.  Added lots of comments, and left some code commented out, mainly to attach a comment to the removed code.  In the next commit, we can delete these commented blocks.

Also renamed the drive subsystem methods related to the gyro to "heading" instead of "angle" as the term is more specific.