Investigated slow drive effort, left comments
Left TODO comments tracing the source of the slow drive (and to a lesser extent, rotation).  The issue is in DriveWithJoystickCommand, but I traced the values through to check for other scale factor problems.  We'll want to edit these files on Saturday to fix them, and remove most of the TODO comments related to scaling.