Calibrated drive motor gear ratio
Calibrated by moving the robot a known distance and reading the number of rotations the motor moved.  The resulting ratio, 6.2, matches one of the options for WCP Swerve2i X2 gear ratio pack.

Odometry reads about 3/4 of the expected distance, even after this change.  But we are having communication trouble with one drive motor, and its rotation value is not updating in the dashboard.  If odometry effectively averages the 4 module delta distances when the robot is traveling in a straight line, then 3/4 is understandable.